refactor: use PascalCase for files having a class (Scenes, Objects)
